[BE] Battery CRUD, Admin CRUD, Station CRUD 검증로직 추가 (#129)
1. 카카오 결제 수정 및 완성
2. data.sql 더미 데이터 추가 및 부분병경 필드값 수정 (reservation, payment, station추가, member detailAddress추가)
3. 배터리 가격구성 변경 (default가격 추가, 더미데이터 가격 변경)
4. 배터리 API에서 할증료 계산 후 반환하는 로직으로 수정
5. Battery CRUD, Admin CRUD, Station CRUD 검증로직 구현
6. Station API : Station 범위&시간 조건검색 수정, Station 시간조건 전체검색 API 구현
7. Member API detailAddress 필드 추가

+++ backend/src/main/java/backend/domain/battery/service/BatteryService.java
@@ -1,5 +1,7 @@
 package backend.domain.battery.service;
 
+import backend.domain.admin.entity.Admin;
+import backend.domain.admin.repository.AdminRepository;
 import backend.domain.battery.entity.Battery;
 import backend.domain.battery.repository.BatteryRepository;
 import backend.domain.station.entity.Station;
@@ -7,34 +9,54 @@ import backend.domain.station.repository.StationRepository;
 import backend.global.exception.dto.BusinessLogicException;
 import backend.global.exception.exceptionCode.ExceptionCode;
 import lombok.RequiredArgsConstructor;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDateTime;
+import java.util.List;
 import java.util.Optional;
 
 @Service //@RequiredArgsConstructor
 public class BatteryService {
     private final BatteryRepository batteryRepository;
     private final StationRepository stationRepository;
+    private final AdminRepository adminRepository;
+    @Value("${mail.address.admin.list}")
+    private List<String> adminMailAddress;
 
-    public BatteryService(BatteryRepository batteryRepository, StationRepository stationRepository){
+    public BatteryService(BatteryRepository batteryRepository, StationRepository stationRepository,  AdminRepository adminRepository){
         this.batteryRepository = batteryRepository;
         this.stationRepository = stationRepository;
+        this.adminRepository = adminRepository;
     }
 
     // 배터리 등록
-    public Battery createBattery(Battery battery, long stationId){
+    public Battery createBattery(Battery battery, long stationId, String adminEmail){
         Station station = stationRepository.findById(stationId)
                 .orElseThrow(() -> new BusinessLogicException(ExceptionCode.STATION_NOT_FOUND));
+
+        // 로그인한 계정이 ADMIN인지 검증하는 로직
+        if(!adminMailAddress.contains(adminEmail)) throw new BusinessLogicException(ExceptionCode.NON_ACCESS_AUTH);
+
+        // 로그인 한 Admin이 실제 그 station의 주인인지 검증
+        verifyAdmin(station.getAdmin().getAdminId(),adminEmail);
+
         battery.setStation(station);
         return batteryRepository.save(battery);
     }
 
     // 배터리 수정
-    public Battery updateBattery(long batteryId){
+    public Battery updateBattery(long batteryId, String adminEmail){
         Battery findBattery = findVerifiedBattery(batteryId);
+
+        // 로그인한 계정이 ADMIN인지 검증하는 로직
+        if(!adminMailAddress.contains(adminEmail)) throw new BusinessLogicException(ExceptionCode.NON_ACCESS_AUTH);
+
+        // 로그인 한 Admin이 실제 그 station의 주인인지 검증
+        verifyAdmin(findBattery.getStation().getAdmin().getAdminId(), adminEmail);
+
         if(!findBattery.isStatus()){
             findBattery.setStatus(true);
         }else{
@@ -59,8 +81,15 @@ public class BatteryService {
     }
 
     // 해당 ID 배터리 삭제
-    public void deleteBattery(long batteryId){
+    public void deleteBattery(long batteryId, String adminEmail){
         Battery findBattery = findVerifiedBattery(batteryId);
+
+        // 로그인한 계정이 ADMIN인지 검증하는 로직
+        if(!adminMailAddress.contains(adminEmail)) throw new BusinessLogicException(ExceptionCode.NON_ACCESS_AUTH);
+
+        // 로그인 한 Admin이 실제 그 station의 주인인지 검증
+        verifyAdmin(findBattery.getStation().getAdmin().getAdminId(), adminEmail);
+
         batteryRepository.delete(findBattery);
     }
 
@@ -71,5 +100,13 @@ public class BatteryService {
         Battery findBattery = optionalBattery.orElseThrow(() ->
                 new BusinessLogicException(ExceptionCode.BATTERY_NOT_FOUND));
         return findBattery;
+    }
+
+    private void verifyAdmin (Long savedAdminId, String adminEmail) {
+        Admin admin = adminRepository.findByEmail(adminEmail)
+                .orElseThrow(() -> new BusinessLogicException(ExceptionCode.ADMIN_NOT_FOUND));
+        if(savedAdminId!=admin.getAdminId()){
+            throw new BusinessLogicException(ExceptionCode.NON_ACCESS_MODIFY);
+        }
     }
 }
